[refactor] Split gpio module into several different modules
This commit takes the functions and definitions inside of the gpio.h
module and splits them into several different modules. The modules
include:
    - [pwm]
    - [adc]
    - [rcc]
    - [tim]

Each is split into a new header file, and a new source file ('.c') was
added for those with methods
