+++ src/startup.c
@@ -1,5 +1,5 @@
 
-#include "armory/gpio.h"
+#include "armory/rcc.h"
 
 int main(void);
